moved models package and get rid of gorm
